+++ src/backends/mod.rs
@@ -8,7 +8,6 @@ pub mod registry;
 
 pub mod appimage;
 pub mod btrfs;
-pub mod conda;
 pub mod emacs;
 pub mod firewall;
 pub mod flatpak;